Challenges: Add new field is_submission_public to Challenge Phase model (#900)
This commit adds a new field `is_submission_public` to the Challenge Phase model. The use of this field is when a phase is public and the host wants to set the default visibility of the submissions being made to it. This way, if the host set it to False(by default), then the participants will have a choice to either show or hide their submissions on leaderboard.