Add check to filter out disabled challenges for create/update challenge
This PR adds a check to the method that creates or updates a GitHub challenge to filter out the disabled challenges so that the hosts are unable to update them. 

This is needed because of a recent issue with duplicate challenges being created on multiple updates of the same challenge.